smartplant-admin-expo/smartplant_iot/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
pyserial>=3.5
geopy>=2.4.0
requests>=2.31.0

